pipeline: skip uploading files after caching fails
It's been brought up that if we're unable to "cache" files in
uploaded/ then we shouldn't try to upload them.

I'm undecided with this change because if we can continue uploading
then the results can be evented, on the audittrail and/or on the
remote server.

However if we cannot upload elsewhere in addition to the cache then
deeper problems should be fixed.